Tests fixed, algorithm for find next line works, sort of.
Tests where giving out inaccurate values becuase the values I comapred
in the beginning where wrong, meaning the source I got it was
wrong. Fixed that. So the some values of the unknown and wanted lines
are correctly calculated like the slope. But the co-ordinates
themselves are wrong after couple of iterations for the same pentagon.