getPolygonArea is implemented but is inaccurate
Because the co-ordinates fed in are generated by my algorithm is
slighly inaccurate becuase it is loosing some decimal points I think.
Which when def into this method may give values upto +-1 unit^2
inaccurate.

The problem with this not working yesterday was C++ returning 1/2 = 0,
should use 1.0/2.0. Also differeve must be added and then subtracted
not subtracted and then substituted