Added note about using which method: Read below
During testing it was found that finding the area using points instead
of themore accurate methods I have implemented before is closer to
the area of the total polygon , which is also found out using the
shoelace formula. Making them as accurate as possible and as close and
possible is the next task.