Deleted unwanted methods from last version of algorithm. Read Below
Unwanted crap from last version is deleted. New tiling methods and
area finding algorithms are in use now in the main
function. Integratation tests for the new version of the algorithm is
also written and shows passing when the parameters are in low scale
with an accuracy of 0.5%. Version 2.0 coming up.